+++ src/ShapeDefinition.php
@@ -16,7 +16,6 @@ final readonly class ShapeDefinition
 	public function __construct(
 		public bool $list,
 		public Extra $extra,
-		public ?string $title,
 		public array $rules,
 		public Contract\ValidatorRegistry $validators,
 		public Contract\CoercerRegistry $coercers,